Update About spoofing Advanced Settings (note) in README.md
A few days ago in telegram groups I saw some people commenting about achieving strong integrity (A13+) by using a revoked keybox, you had to set spoofProvider to 1 and you effectively achieved strong integrity, this false integrity will probably be temporary until the keybox becomes completely useless, but it seems that a valid keybox (a leaked one) that I found also needs spoofProvider to 1 to achieve strong. I don't know if this change is also necessary for private keyboxes or if this change will be the same for all leaked keyboxes from now on. I don't have another phone to check if this change also applies to <A13 devices, it would be nice if another user could help with this verification.